+++ hub/src/main/java/com/zgzg/hub/application/route/service/facade/RouteService.java
@@ -6,7 +6,7 @@ import com.zgzg.common.exception.BaseException;
 import com.zgzg.hub.application.route.dto.RedisRouteDTO;
 import com.zgzg.hub.application.route.dto.RoutesDTO;
 import com.zgzg.hub.domain.entity.Route;
-import com.zgzg.hub.domain.repository.RouteRepository;
+import com.zgzg.hub.domain.repository.route.RouteRepository;
 import com.zgzg.hub.infrastructure.redis.RedisHashUtil;
 import java.util.HashMap;
 import java.util.List;
